requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
PyMuPDF==1.21.1
PyPDF2==2.10.8
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
